# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MicroLink Control System \u2014 Data Platform API",
    "description": "**Stream B Data Platform API** \u2014 the central nervous system of MCS.\n\nThis is the contract that Stream C (Business Systems) and Stream D (Frontend)\ndevelop against. All telemetry, alarm, billing, and real-time data flows through\nthese endpoints.\n\n## Authentication\nAPI key via `X-API-Key` header. Multi-tenant isolation via row-level security\nscoped to the authenticated tenant.\n\n## Versioning\nAll endpoints are prefixed with `/api/v1`. Breaking changes will increment\nthe version number.\n\n## Rate Limits\n- REST endpoints: 1,000 req/min per API key\n- WebSocket connections: 10 per API key\n- Telemetry queries spanning >90 days: 10 req/min\n\n## Real-Time Feeds\nWebSocket endpoints provide live data streams backed by Redis pub/sub.\nReconnect with exponential backoff on disconnect.\n",
    "version": "0.1.0",
    "contact": {
      "name": "MicroLink MCS Team",
      "email": "engineering@microlinkdc.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://api.mcs.microlinkdc.com/api/v1",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/alarms": {
      "get": {
        "tags": [
          "Alarms"
        ],
        "summary": "List alarms",
        "description": "Returns non-CLEARED alarms sorted by priority then time.",
        "operationId": "listAlarms",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated: ACTIVE,ACKED,RTN_UNACK,SHELVED,SUPPRESSED",
            "example": "ACTIVE,RTN_UNACK"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated: P0,P1,P2,P3",
            "example": "P0,P1"
          },
          {
            "name": "block_slug",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "site_slug",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/alarms/{sensor_id}/acknowledge": {
      "post": {
        "tags": [
          "Alarms"
        ],
        "summary": "Acknowledge an alarm",
        "description": "Transitions ACTIVE \u2192 ACKED or RTN_UNACK \u2192 CLEARED.\nLogs an immutable event with the operator identity.\n",
        "operationId": "acknowledgeAlarm",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlarmAckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alarm"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/alarms/{sensor_id}/shelve": {
      "post": {
        "tags": [
          "Alarms"
        ],
        "summary": "Shelve an alarm",
        "description": "Temporarily suppress an alarm. Requires a reason (audit trail).\nMaximum duration: 24 hours. Timer-expired alarms auto-unshelve\nand re-evaluate.\n",
        "operationId": "shelveAlarm",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlarmShelveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alarm"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/alarms/stats": {
      "get": {
        "tags": [
          "Alarms"
        ],
        "summary": "ISA-18.2 compliance stats",
        "operationId": "alarmStats",
        "parameters": [
          {
            "name": "block_slug",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmStats"
                }
              }
            }
          }
        }
      }
    },
...
    "/billing/kwh": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Electrical energy data (5-min)",
        "description": "5-minute electrical energy (kWh) data for billing.\nStream C multiplies `kwh_value` by customer rate for invoice line items.\n",
        "operationId": "billingKwh",
        "parameters": [
          {
            "name": "block_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingKwhResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/kwht": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Thermal energy data (5-min)",
        "description": "5-minute thermal energy (kWht) data for heat credit billing.\nStream C uses this for host heat recovery credit calculations.\n",
        "operationId": "billingKwht",
        "parameters": [
          {
            "name": "block_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingKwhtResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/energy-daily": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Daily energy summary",
        "description": "Daily energy totals per block \u2014 electrical kWh, thermal recovery,\nPUE estimate. Used by lender reports and ESG module.\n",
        "operationId": "energyDailySummary",
        "parameters": [
          {
            "name": "block_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnergyDailySummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 100,
          "minimum": 1,
          "maximum": 1000
        }
      },
      "Start": {
        "name": "start",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "Start time (ISO 8601)"
      },
      "End": {
        "name": "end",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "End time (ISO 8601, default now)"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Alarm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "nullable": true
          },
          "sensor_id": {
            "type": "integer"
          },
          "priority": {
            "type": "string",
            "enum": [
              "P0",
              "P1",
              "P2",
              "P3"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ACKED",
              "RTN_UNACK",
              "CLEARED",
              "SHELVED",
              "SUPPRESSED"
            ]
          },
          "tag": {
            "type": "string"
          },
          "subsystem": {
            "type": "string"
          },
          "site_id": {
            "type": "string"
          },
          "block_id": {
            "type": "string"
          },
          "value_at_raise": {
            "type": "number",
            "nullable": true
          },
          "value_at_clear": {
            "type": "number",
            "nullable": true
          },
          "threshold_value": {
            "type": "number",
            "nullable": true
          },
          "threshold_direction": {
            "type": "string",
            "nullable": true,
            "enum": [
              "HIGH",
              "LOW"
            ]
          },
          "raised_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "acked_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "acked_by": {
            "type": "string",
            "nullable": true
          },
          "cleared_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shelved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shelved_by": {
            "type": "string",
            "nullable": true
          },
          "shelved_until": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shelve_reason": {
            "type": "string",
            "nullable": true
          },
          "suppressed_by_alarm_id": {
            "type": "integer",
            "nullable": true
          },
          "response_time_seconds": {
            "type": "number",
            "nullable": true
          },
          "last_value": {
            "type": "number",
            "nullable": true
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "AlarmListResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total matching alarms"
          },
          "standing": {
            "type": "integer",
            "description": "ACTIVE + RTN_UNACK count"
          },
          "alarms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alarm"
            }
          }
        }
      },
      "AlarmAckRequest": {
        "type": "object",
        "required": [
          "operator"
        ],
        "properties": {
          "operator": {
            "type": "string",
            "minLength": 1,
            "example": "nick.searra"
          }
        }
      },
      "AlarmShelveRequest": {
        "type": "object",
        "required": [
          "operator",
          "reason"
        ],
        "properties": {
          "operator": {
            "type": "string",
            "minLength": 1
          },
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500,
            "example": "Sensor calibration in progress"
          },
          "duration_hours": {
            "type": "number",
            "default": 8.0,
            "minimum": 0.5,
            "maximum": 24.0
          }
        }
      },
      "AlarmStats": {
        "type": "object",
        "properties": {
          "standing": {
            "type": "integer"
          },
          "acked": {
            "type": "integer"
          },
          "shelved": {
            "type": "integer"
          },
          "suppressed": {
            "type": "integer"
          },
          "raised_last_hour": {
            "type": "integer"
          },
          "avg_response_seconds_24h": {
            "type": "number"
          },
          "isa_18_2_target_per_hour": {
            "type": "integer",
            "example": 6
          },
          "compliant": {
            "type": "boolean"
          }
        }
      },
...
      "BillingKwhRow": {
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string",
            "format": "date-time"
          },
          "sensor_id": {
            "type": "integer"
          },
          "sensor_tag": {
            "type": "string",
            "example": "P-MSB-TOTAL"
          },
          "equipment_tag": {
            "type": "string"
          },
          "block_slug": {
            "type": "string"
          },
          "site_slug": {
            "type": "string"
          },
          "kwh_value": {
            "type": "number",
            "nullable": true,
            "description": "Energy in kWh for this 5-min interval"
          },
          "unit": {
            "type": "string"
          },
          "sample_count": {
            "type": "integer"
          },
          "quality_good_ratio": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "BillingKwhResponse": {
        "type": "object",
        "properties": {
          "block_slug": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "total_kwh": {
            "type": "number",
            "description": "Sum of all kwh_value rows"
          },
          "row_count": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingKwhRow"
            }
          }
        }
      },
      "BillingKwhtRow": {
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string",
            "format": "date-time"
          },
          "sensor_id": {
            "type": "integer"
          },
          "sensor_tag": {
            "type": "string",
            "example": "TM-01-HEAT-RATE"
          },
          "equipment_tag": {
            "type": "string"
          },
          "block_slug": {
            "type": "string"
          },
          "site_slug": {
            "type": "string"
          },
          "kwht_value": {
            "type": "number",
            "nullable": true,
            "description": "Thermal energy in kWht for this 5-min interval"
          },
          "unit": {
            "type": "string"
          },
          "sample_count": {
            "type": "integer"
          },
          "quality_good_ratio": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "BillingKwhtResponse": {
        "type": "object",
        "properties": {
          "block_slug": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "total_kwht": {
            "type": "number"
          },
          "row_count": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingKwhtRow"
            }
          }
        }
      },
      "EnergyDailySummary": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "format": "date-time"
          },
          "block_slug": {
            "type": "string"
          },
          "site_slug": {
            "type": "string"
          },
          "electrical_kwh": {
            "type": "number"
          },
          "thermal_kwht_recovered": {
            "type": "number"
          },
          "thermal_kwht_rejected": {
            "type": "number"
          },
          "heat_recovery_ratio": {
            "type": "number",
            "description": "thermal_recovered / electrical (0.0\u20131.0)"
          },
          "pue_estimate": {
            "type": "number",
            "nullable": true
          }
        }
      },
...
    }
  }
}